requirements.txt
torch>=2.0.0
torchvision>=0.15.0
google-cloud-storage>=2.10.0
opencv-python-headless>=4.8.0
PyYAML>=6.0
mlflow>=2.8.0
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
